maptr/pytorch/requirements.nodeps.txt
mmcv-full==1.7.2
mmdet==2.28.2
yapf==0.43.0   # required for mmcv
